WP 4.2: consolidate spec/requests/extensions into three files
Eight files and 779 lines become three and roughly 380. The setup pattern
"create a themed site, attach the ward" appeared fourteen times; one shared
themed site and one shared core site replace it.

engine_spec.rb is the old example_theme_spec.rb minus the eager-load example
(the constants already resolve in the example above it) and the duplicate
"renders the homepage view for a site".

theme_slots_spec.rb absorbs theme_homepage_spec, theme_head_spec,
theme_icons_spec and day_strip_filter_spec. One positive example per slot,
and one combined negative that asserts a core site renders none of them,
which catches a slot leaking into core regardless of which slot it is.

theme_fallbacks_spec.rb is theme_asset_fallback_spec plus the
unregistered-theme cases that were scattered through the icon and head specs.

The map style example was not a request spec at all, so it moves to
spec/helpers/map_helper_spec.rb.
